+++ UnifiedPatientManager/src/main/java/com/Eges411Team/UnifiedPatientManager/repositories/MedicalHistoryRepo.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Date;
 
-public interface MedicalHistoryRepo extends JpaRepository<MedicalHistory, Integer> {
+public interface MedicalHistoryRepo extends JpaRepository<MedicalHistory, Long> {
 
     // Fetch all medical history records for a given patient
     List<MedicalHistory> findAllByPatient_id(Long patient_id);
